homework_api/requirements.txt
Flask==2.3.2
python-dotenv==1.0.0
requests==2.31.0
Jinja2==3.1.2
opencv-python==4.9.0.80
blinker==1.7.0
click==8.1.7
itsdangerous==2.1.2
MarkupSafe==2.1.5
Werkzeug==3.0.1